Restructure commands
