+++ pom.xml
@@ -12,7 +12,7 @@
     <name>PrisonGame</name>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
