+++ src/main/java/prisongame/prisongame/FilteredWords.java
@@ -1,7 +1,5 @@
 package prisongame.prisongame;
 
-import com.sun.tools.javac.util.StringUtils;
-
 public class FilteredWords {
     static String[] filter = {
             "nigger",
@@ -12,7 +10,7 @@ public class FilteredWords {
             "retard"
     };
 
-    static String filtermsg(String msg) {
+    public static String filtermsg(String msg) {
         for (String i : filter) {
             if (msg.toLowerCase().contains(i)) {
                 msg = "I FUCKING LOVE AMONG US!!! YESS!!! AMONGER!! SUSS!!! SUSSY!!! SUSSY BAKA!! SUSS!! WALTUH!! KINDA SUS WALTUH!!";
