+++ src/main/java/prisongame/prisongame/PrisonGame.java
@@ -22,6 +22,16 @@ import org.bukkit.potion.PotionEffectType;
 import org.bukkit.scoreboard.Team;
 import org.bukkit.util.Vector;
 import oshi.jna.platform.mac.SystemB;
+import prisongame.prisongame.commands.*;
+import prisongame.prisongame.commands.danger.HardCommand;
+import prisongame.prisongame.commands.danger.NormalCommand;
+import prisongame.prisongame.commands.danger.ResetAscensionCommand;
+import prisongame.prisongame.commands.economy.PayCommand;
+import prisongame.prisongame.commands.economy.staff.NerdCheatCommand;
+import prisongame.prisongame.commands.economy.staff.ResetMoneyCommand;
+import prisongame.prisongame.commands.economy.staff.SetMoneyCommand;
+import prisongame.prisongame.commands.staff.BuilderCommand;
+import prisongame.prisongame.commands.warden.ResignCommand;
 
 import java.io.Console;
 import java.util.Arrays;
@@ -35,16 +45,16 @@ public final class PrisonGame extends JavaPlugin {
 
     static HashMap<Player, Double> st = new HashMap<>();
     static HashMap<Player, Double> sp = new HashMap<>();
-    static Player warden = null;
-    static HashMap<Player, Boolean> escaped = new HashMap<>();
-    static HashMap<Player, Integer> type = new HashMap<>();
-    static HashMap<Player, Integer> askType = new HashMap<>();
+    public static Player warden = null;
+    public static HashMap<Player, Boolean> escaped = new HashMap<>();
+    public static HashMap<Player, Integer> type = new HashMap<>();
+    public static HashMap<Player, Integer> askType = new HashMap<>();
     static HashMap<Player, Integer> lastward = new HashMap<>();
     static HashMap<Player, Integer> lastward2 = new HashMap<>();
     static HashMap<Player, Integer> wardenban = new HashMap<>();
     static HashMap<Player, String> word = new HashMap<>();
     static HashMap<Player, Integer> saidcycle = new HashMap<>();
-    static HashMap<Player, String> prisonnumber = new HashMap<>();
+    public static HashMap<Player, String> prisonnumber = new HashMap<>();
     static HashMap<Player, Double> wealthcycle = new HashMap<>();
     static HashMap<Player, Integer> wardentime = new HashMap<>();
     static HashMap<Player, Integer> calls = new HashMap<>();
@@ -63,39 +73,39 @@ public final class PrisonGame extends JavaPlugin {
     static Prison boat;
     static Prison nether;
     static Prison amongus;
-    static Integer solitcooldown = 0;
-    static Prison active = null;
+    public static Integer solitcooldown = 0;
+    public static Prison active = null;
     static NamespacedKey nightvis;
     static NamespacedKey rank;
     static NamespacedKey coarsemined;
     static NamespacedKey hg;
-    static NamespacedKey ascendcoins;
+    public static NamespacedKey ascendcoins;
     static NamespacedKey doubincome;
-    static NamespacedKey taxevasion;
+    public static NamespacedKey taxevasion;
     static NamespacedKey rankprefix;
-    static NamespacedKey semicloak;
-    static NamespacedKey reinforcement;
-    static NamespacedKey bckupmny;
-    static NamespacedKey protspawn;
+    public static NamespacedKey semicloak;
+    public static NamespacedKey reinforcement;
+    public static NamespacedKey bckupmny;
+    public static NamespacedKey protspawn;
     static NamespacedKey tab;
-    static NamespacedKey randomz;
+    public static NamespacedKey randomz;
     static Integer swapcool = 0;
-    static Integer wardenCooldown = 20;
+    public static Integer wardenCooldown = 20;
     static Integer lockdowncool = 0;
     static NamespacedKey whiff;
-    static Boolean wardenenabled = false;
+    public static Boolean wardenenabled = false;
 
-    static NamespacedKey mny;
+    public static NamespacedKey mny;
     static HashMap<Player, Integer> respect = new HashMap<>();
-    static HashMap<Player, Integer> solittime = new HashMap<>();
+    public static HashMap<Player, Integer> solittime = new HashMap<>();
     static HashMap<Material, Double> moneyore = new HashMap<>();
     static HashMap<Player, Player> handcuff = new HashMap<>();
     static HashMap<Player, Integer> trustlevel = new HashMap<>();
     static HashMap<Player, Integer> prisonerlevel = new HashMap<>();
     static NamespacedKey trust;
-    static HashMap<Player, Boolean> hardmode = new HashMap<>();
+    public static HashMap<Player, Boolean> hardmode = new HashMap<>();
     static HashMap<Player, Player> killior = new HashMap<>();
-    static NamespacedKey muted;
+    public static NamespacedKey muted;
 
     static Material[] oretypes = {
             Material.DEEPSLATE_COPPER_ORE,
@@ -114,7 +124,7 @@ public final class PrisonGame extends JavaPlugin {
     static Villager bmsh1;
     static Villager bmsh2;
     static Villager shop;
-    static Boolean swat = false;
+    public static Boolean swat = false;
 
     public static void tptoBed(Player p) {
         p.teleport(PrisonGame.active.getNursebedOutTP());
@@ -205,26 +215,21 @@ public final class PrisonGame extends JavaPlugin {
         moneyore.put(Material.DEEPSLATE_REDSTONE_ORE, 10.0);
 
         Bukkit.broadcastMessage("RELOAD: Loaded NameSpacedKeys");
-        this.getCommand("warden").setExecutor(new CommandKit());
-        this.getCommand("resign").setExecutor(new TestCommand());
-        this.getCommand("tc").setExecutor(new TeamChat());
-        this.getCommand("disc").setExecutor(new Discordcmd());
-        this.getCommand("accept").setExecutor(new accpt());
-        this.getCommand("hello").setExecutor(new hello());
-        this.getCommand("nerdcheatcommand").setExecutor(new shittonmoney());
-        //this.getCommand("bragrightpackage").setExecutor(new bragright());
-        this.getCommand("rstmoney").setExecutor(new nomone());
-        this.getCommand("amute").setExecutor(new accmute());
-        this.getCommand("aunmute").setExecutor(new accunmute());
-        this.getCommand("pay").setExecutor(new gib());
-        this.getCommand("safereload").setExecutor(new sfreload());
-        this.getCommand("builder").setExecutor(new BuilderCMD());
-        this.getCommand("rstascen").setExecutor(new resetasc());
-        this.getCommand("hard").setExecutor(new hard());
-        this.getCommand("normal").setExecutor(new normal());
-        this.getCommand("setmoney").setExecutor(new stmoney());
-        this.getCommand("pbsettings").setExecutor(new settings());
-        this.getCommand("giverank").setExecutor(new giverank());
+        this.getCommand("warden").setExecutor(new WardenCommand());
+        this.getCommand("resign").setExecutor(new ResignCommand());
+        this.getCommand("tc").setExecutor(new TeamChatCommand());
+        this.getCommand("disc").setExecutor(new DiscCommand());
+        this.getCommand("accept").setExecutor(new AcceptCommand());
+        this.getCommand("hello").setExecutor(new HelloCommand());
+        this.getCommand("nerdcheatcommand").setExecutor(new NerdCheatCommand());
+        this.getCommand("rstmoney").setExecutor(new ResetMoneyCommand());
+        this.getCommand("pay").setExecutor(new PayCommand());
+        this.getCommand("builder").setExecutor(new BuilderCommand());
+        this.getCommand("rstascen").setExecutor(new ResetAscensionCommand());
+        this.getCommand("hard").setExecutor(new HardCommand());
+        this.getCommand("normal").setExecutor(new NormalCommand());
+        this.getCommand("setmoney").setExecutor(new SetMoneyCommand());
+        this.getCommand("pbsettings").setExecutor(new PBSettingsCommand());
 
         for (Player p : Bukkit.getOnlinePlayers()) {
             p.sendMessage("RELOAD: Loaded Commands");
@@ -284,7 +289,7 @@ public final class PrisonGame extends JavaPlugin {
             Bukkit.broadcastMessage("RELOAD: Removed Entities");
             gaeae = new Prison("Gaeae Fort", new Location(Bukkit.getWorld("world"), 61, -54, -159), new Location(Bukkit.getWorld("world"), 76, -59, -169), new Location(Bukkit.getWorld("world"), 44, -58, -141), new Location(Bukkit.getWorld("world"), 44, -58, -137), new Location(Bukkit.getWorld("world"), 41.5, -52, -120.5), new Location(Bukkit.getWorld("world"), 12, -60, -119), new Location(Bukkit.getWorld("world"), -26.5, -56.5, -115.5), new Location(Bukkit.getWorld("world"), -8.5, -57, -108.5), new Location(Bukkit.getWorld("world"), 33, -59, -132), new Location(Bukkit.getWorld("world"), 70.5, -59, -137.5), new Location(Bukkit.getWorld("world"), 87, -59, -129), new Location(Bukkit.getWorld("world"), 87, -56, -125));
             hyper = new Prison("HyperTech", new Location(Bukkit.getWorld("world"), 18, -56, -988), new Location(Bukkit.getWorld("world"), 8, -59, -981), new Location(Bukkit.getWorld("world"), -29, -58, -988, 0, 0), new Location(Bukkit.getWorld("world"), -29, -58, -991), new Location(Bukkit.getWorld("world"), 12, -53, -970), new Location(Bukkit.getWorld("world"), -18, -59, -995), new Location(Bukkit.getWorld("world"), -35.5, -56.5, -959.5), new Location(Bukkit.getWorld("world"), 3.5, -59, -1006.5), new Location(Bukkit.getWorld("world"), 13, -59, -1009), new Location(Bukkit.getWorld("world"), 0.5, -59, -996.5), new Location(Bukkit.getWorld("world"), 1, -58, -1008), new Location(Bukkit.getWorld("world"), 3, -58, -1008));
-            endmap = new Prison("The End?", new Location(Bukkit.getWorld("endprison"), 7, 133, 8), new Location(Bukkit.getWorld("endprison"), 19, 127, 20), new Location(Bukkit.getWorld("endprison"), -30, 170, -48), new Location(Bukkit.getWorld("endprison"), -32, 170, -47), new Location(Bukkit.getWorld("endprison"), -1, 150, 13), new Location(Bukkit.getWorld("endprison"), 0, 135, -41), new Location(Bukkit.getWorld("world"), -26.5, -56.5, -115.5), new Location(Bukkit.getWorld("endprison"), 0, 131, -6), new Location(Bukkit.getWorld("endprison"), 0, 125 ,0), new Location(Bukkit.getWorld("endprison"), 16.5, 129, 2.5), new Location(Bukkit.getWorld("endprison"), -100000, 256, -100000), new Location(Bukkit.getWorld("endprison"), -100000, 256, -100000));
+            //endmap = new Prison("The End?", new Location(Bukkit.getWorld("endprison"), 7, 133, 8), new Location(Bukkit.getWorld("endprison"), 19, 127, 20), new Location(Bukkit.getWorld("endprison"), -30, 170, -48), new Location(Bukkit.getWorld("endprison"), -32, 170, -47), new Location(Bukkit.getWorld("endprison"), -1, 150, 13), new Location(Bukkit.getWorld("endprison"), 0, 135, -41), new Location(Bukkit.getWorld("world"), -26.5, -56.5, -115.5), new Location(Bukkit.getWorld("endprison"), 0, 131, -6), new Location(Bukkit.getWorld("endprison"), 0, 125 ,0), new Location(Bukkit.getWorld("endprison"), 16.5, 129, 2.5), new Location(Bukkit.getWorld("endprison"), -100000, 256, -100000), new Location(Bukkit.getWorld("endprison"), -100000, 256, -100000));
             train = new Prison("Train", nl("world",0D,0D,0D, 0f, 0f), nl("world",0D,0D,0D, 0f, 0f), nl("world",82.8D,-58D,953D, 0f, 0f), nl("world",82.8D,-58D,953D, 0f, 0f), nl("world",83.0D,-54D,980D, -180f, 0f), nl("world",91.9D,-58D,982D, 0f, 0f), nl("world",64D,-56D,981D, -180f, 0f), nl("world",75.5D,-58D,964.5D, 0f, 0f), nl("world",91D,-58D,946D, 0f, 0f), nl("world",80.5D,-54D,961.5D, 0f, 0f), nl("world",79D,-58D,976D, 0f, 0f), nl("world",79D,-56D,975D, 0f, 0f));
             gladiator = new Prison("Gladiator", new Location(Bukkit.getWorld("world"), -2024, -55, 1919), new Location(Bukkit.getWorld("world"), -1999, -60, 1940), new Location(Bukkit.getWorld("world"), -2039, -59, 1933, 180, 0), new Location(Bukkit.getWorld("world"), -2040, -59, 1933, 180, 0), new Location(Bukkit.getWorld("world"), -2041, -47, 1957), new Location(Bukkit.getWorld("world"), -2037, -60, 1947), new Location(Bukkit.getWorld("world"), -2030, -60, 2015), new Location(Bukkit.getWorld("world"), -2022, -60, 1967), new Location(Bukkit.getWorld("world"), -1973, -60, 1984), new Location(Bukkit.getWorld("world"), -2022.5, -60, 1957.5), new Location(Bukkit.getWorld("world"), -2084, -60, 1973), new Location(Bukkit.getWorld("world"), -2080, -56, 1973));
             island = new Prison("Island", new Location(Bukkit.getWorld("world"), 1976, -55, -2001), new Location(Bukkit.getWorld("world"), 1968, -60, -2009), new Location(Bukkit.getWorld("world"), 2003, -59, -1988, 90, 0), new Location(Bukkit.getWorld("world"), 2003, -59, -1988, 90, 0), new Location(Bukkit.getWorld("world"), 1988, -60, -1978), new Location(Bukkit.getWorld("world"), 1964, -60, -1981), new Location(Bukkit.getWorld("world"), 1990, -59, -1861), new Location(Bukkit.getWorld("world"), 1989, -60, -1990), new Location(Bukkit.getWorld("world"), 1979, -60, -1982), new Location(Bukkit.getWorld("world"), 1981.5, -60, -1989.5), new Location(Bukkit.getWorld("world"), 1958, -60, -1999), new Location(Bukkit.getWorld("world"), 1962, -57, -1999));
@@ -410,7 +415,7 @@ public final class PrisonGame extends JavaPlugin {
     static Location nl(String world, Double X, Double Y, Double Z, Float yaw, Float pitch) {
         return new Location(Bukkit.getWorld(world), X, Y, Z, yaw, pitch);
     }
-    static void setNurse(Player g) {
+    public static void setNurse(Player g) {
         Bukkit.getScoreboardManager().getMainScoreboard().getTeam("Guards").addPlayer(g);
         PrisonGame.type.put(g, 2);
         Bukkit.broadcastMessage(ChatColor.LIGHT_PURPLE + g.getName() + " was promoted to a nurse!");
@@ -488,7 +493,7 @@ public final class PrisonGame extends JavaPlugin {
         }
 
     }
-    static void setSwat(Player g) {
+    public static void setSwat(Player g) {
         Bukkit.getScoreboardManager().getMainScoreboard().getTeam("Guards").addPlayer(g);
         PrisonGame.type.put(g, 3);
         Bukkit.broadcastMessage(ChatColor.DARK_GRAY + g.getName() + " was promoted to a SWAT member!");
@@ -553,7 +558,7 @@ public final class PrisonGame extends JavaPlugin {
 
     }
 
-    static void setGuard(Player g) {
+    public static void setGuard(Player g) {
         Bukkit.getScoreboardManager().getMainScoreboard().getTeam("Guards").addPlayer(g);
         PrisonGame.type.put(g, 1);
         Bukkit.broadcastMessage(ChatColor.BLUE + g.getName() + " was promoted to a guard!");
@@ -662,414 +667,3 @@ public final class PrisonGame extends JavaPlugin {
     }
 
 }
-
-class TestCommand implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        if (!((Player) sender).hasCooldown(Material.IRON_DOOR)) {
-            if (!((Player) sender).getDisplayName().contains("SOLITARY")) {
-                if (PrisonGame.warden != null) {
-                    if (PrisonGame.warden.equals(sender)) {
-                        PrisonGame.wardenCooldown = 20 * 3;
-                        PrisonGame.warden = null;
-                    }
-                }
-                PrisonGame.type.put((Player) sender, 0);
-                MyListener.playerJoin((Player) sender, false);
-            }
-        } else {
-            sender.sendMessage(ChatColor.RED + "You're in combat!");
-        }
-        return true;
-    }
-}
-
-class accmute implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        if (args[0] == null) {
-            sender.sendMessage("you didnt add a player idiot fucking stupid!!!");
-            return true;
-        }
-        if (args[1] == null) {
-            sender.sendMessage("add a time (in seconds) too!!! smh dumbasss!! idiot!!!!!!!");
-            return true;
-        }
-        Bukkit.getPlayer(args[0]).getPersistentDataContainer().set(PrisonGame.muted, PersistentDataType.INTEGER, Integer.valueOf(args[1]) * 20);
-        Bukkit.broadcastMessage(ChatColor.GREEN + args[0] + " was muted for" + Integer.valueOf(args[1]) * 20 + " seconds!!!! lmao laugh at this user!!");
-        return true;
-    }
-}
-
-class accunmute implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        Bukkit.getPlayer(args[0]).getPersistentDataContainer().remove(PrisonGame.muted);
-        return true;
-    }
-}
-
-class bragright implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        if (!(sender instanceof Player)) {
-            Bukkit.getPlayer(args[0]).getPersistentDataContainer().set(PrisonGame.rank, PersistentDataType.INTEGER, 1);
-        }
-        return true;
-    }
-}
-
-
-class Discordcmd implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        sender.sendMessage(ChatColor.BLUE + "https://discord.gg/GrcHKkFQsv");
-        return true;
-    }
-}
-
-
-    class hello implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        sender.sendMessage(ChatColor.GRAY + "Hello! " + ChatColor.GOLD + "You're currently playing on " + ChatColor.BLUE + "PrisonButBad.minehut.gg" + ChatColor.RED + ", You're on the " + ChatColor.WHITE + PrisonGame.active.name + " map, " + ChatColor.DARK_GREEN + " with " + Bukkit.getOnlinePlayers().size() + "/" + Bukkit.getMaxPlayers() + " players online. " + ChatColor.GRAY + "(PrisonButBad made by agmass)");
-        return true;
-    }
-}
-
-class shittonmoney implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        Bukkit.broadcastMessage("lmao!!! " + sender.getName() + " used the cheat command to give them 1000$!! probbably was just testing but what a pussy L!!!!");
-        Player p = (Player) sender;
-        p.getPersistentDataContainer().set(PrisonGame.mny, PersistentDataType.DOUBLE ,p.getPersistentDataContainer().getOrDefault(PrisonGame.mny, PersistentDataType.DOUBLE, 0.0)+ 1000.0);
-        return true;
-    }
-}
-
-class stmoney implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        Player p = Bukkit.getPlayer(args[0]);
-        p.getPersistentDataContainer().set(PrisonGame.mny, PersistentDataType.DOUBLE ,Double.valueOf(args[1]));
-        return true;
-    }
-}
-
-class sttrust implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        Player p = Bukkit.getPlayer(args[0]);
-        p.getPersistentDataContainer().set(PrisonGame.trust, PersistentDataType.DOUBLE ,Double.valueOf(args[1]));
-        return true;
-    }
-}
-
-
-class nomone implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        Bukkit.getPlayer(args[0]).getPersistentDataContainer().set(PrisonGame.mny, PersistentDataType.DOUBLE ,0.0);
-        return true;
-    }
-}
-
-class giverank implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        if (args.length == 1) {
-            Bukkit.getPlayer(args[0]).getPersistentDataContainer().remove(PrisonGame.rankprefix);
-        } else {
-            Bukkit.getPlayer(args[0]).getPersistentDataContainer().set(PrisonGame.rankprefix, PersistentDataType.STRING, args[1]);
-        }
-        return true;
-    }
-}
-
-class resetasc implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        if (sender instanceof Player) {
-            Player p = (Player) sender;
-            p.getPersistentDataContainer().remove(PrisonGame.randomz);
-            p.getPersistentDataContainer().remove(PrisonGame.taxevasion);
-            p.getPersistentDataContainer().remove(PrisonGame.protspawn);
-            p.getPersistentDataContainer().remove(PrisonGame.reinforcement);
-            p.getPersistentDataContainer().remove(PrisonGame.semicloak);
-            p.getPersistentDataContainer().remove(PrisonGame.ascendcoins);
-            p.sendMessage(ChatColor.RED + "Your ascension has been reset!");
-        }
-        return true;
-    }
-}
-
-
-class gib implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        Player p = (Player) sender;
-        if (Double.valueOf(args[1]) > 0) {
-            if (p.getPersistentDataContainer().get(PrisonGame.mny, PersistentDataType.DOUBLE) >= Double.valueOf(args[1])) {
-                Bukkit.getPlayer(args[0]).getPersistentDataContainer().set(PrisonGame.mny, PersistentDataType.DOUBLE, Bukkit.getPlayer(args[0]).getPersistentDataContainer().get(PrisonGame.mny, PersistentDataType.DOUBLE) + Double.valueOf(args[1]));
-                p.getPersistentDataContainer().set(PrisonGame.mny, PersistentDataType.DOUBLE, p.getPersistentDataContainer().get(PrisonGame.mny, PersistentDataType.DOUBLE) - Double.valueOf(args[1]));
-            }
-        }
-        return true;
-    }
-}
-
-class hard implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        if (sender instanceof Player && !PrisonGame.hardmode.get((Player) sender)) {
-            if (!((Player) sender).getDisplayName().contains("SOLITARY") && !((Player) sender).hasCooldown(Material.IRON_DOOR)) {
-                Player p = (Player) sender;
-                p.setViewDistance(2);
-                p.getPersistentDataContainer().set(PrisonGame.bckupmny, PersistentDataType.DOUBLE, p.getPersistentDataContainer().getOrDefault(PrisonGame.mny, PersistentDataType.DOUBLE, 0.0));
-                p.getPersistentDataContainer().set(PrisonGame.mny, PersistentDataType.DOUBLE, 0.0);
-                if (PrisonGame.warden != null) {
-                    if (PrisonGame.warden.equals(sender)) {
-                        PrisonGame.warden = null;
-                    }
-                }
-                PrisonGame.type.put((Player) sender, 0);
-                PrisonGame.hardmode.put(p, true);
-                MyListener.playerJoin(p, false);
-                String prisonerNumber = "" + new Random().nextInt(100, 999);
-                PrisonGame.prisonnumber.put(p, prisonerNumber);
-                PlayerDisguise playerDisguise = new PlayerDisguise("pdlCAMERA");
-                playerDisguise.setName("Prisoner " + prisonerNumber);
-                playerDisguise.setKeepDisguiseOnPlayerDeath(true);
-                DisguiseAPI.disguiseToAll(p, playerDisguise);
-                p.setCustomName(ChatColor.GRAY + "[" + ChatColor.GOLD + "PRISONER" + ChatColor.GRAY + "] " + ChatColor.DARK_GRAY  + "Prisoner " + prisonerNumber);
-                p.setDisplayName(ChatColor.GRAY + "[" + ChatColor.GOLD + "PRISONER" + ChatColor.GRAY + "] " + ChatColor.DARK_GRAY  + "Prisoner " + prisonerNumber);
-
-            }
-        }  else {
-            sender.sendMessage(ChatColor.RED + "Really? Did you deadass try to run hard when already in hard mode? Come on. You're better than this. (I had to add this since people were losing money by doing hard... come on...)");
-        }
-        return true;
-    }
-}
-
-class normal implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        if (sender instanceof Player && PrisonGame.hardmode.get((Player) sender)) {
-            if (!((Player) sender).getDisplayName().contains("SOLITARY") && !((Player) sender).hasCooldown(Material.IRON_DOOR)) {
-                Player p = (Player) sender;
-                p.getPersistentDataContainer().set(PrisonGame.mny, PersistentDataType.DOUBLE, p.getPersistentDataContainer().getOrDefault(PrisonGame.bckupmny, PersistentDataType.DOUBLE, 0.0));
-                p.getPersistentDataContainer().set(PrisonGame.bckupmny, PersistentDataType.DOUBLE, 0.0);
-                if (PrisonGame.warden != null) {
-                    if (PrisonGame.warden.equals(sender)) {
-                        PrisonGame.warden = null;
-                    }
-                }
-                PrisonGame.type.put((Player) sender, 0);
-                DisguiseAPI.undisguiseToAll(p);
-                PrisonGame.hardmode.put(p, false);
-                MyListener.playerJoin(p, false);
-            }
-        } else {
-            sender.sendMessage(ChatColor.RED + "Really? Did you deadass try to run normal when already in normal mode? Come on. You're better than this. (I had to add this since people were losing money by doing normal... come on...)");
-        }
-        return true;
-    }
-}
-
-class accpt implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-            if (PrisonGame.askType.getOrDefault((Player) sender, 0) == 2) {
-                PrisonGame.setNurse((Player) sender);
-                Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "advancement grant " + sender.getName() + " only prison:guard");
-                Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "advancement grant " + PrisonGame.warden.getName() + " only prison:support");
-                Player nw = (Player) sender;
-                if (nw.getPersistentDataContainer().has(PrisonGame.protspawn, PersistentDataType.INTEGER)) {
-                    if (nw.getInventory().getHelmet() != null)
-                        nw.getInventory().getHelmet().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
-                    if (nw.getInventory().getChestplate() != null)
-                        nw.getInventory().getChestplate().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
-                    if (nw.getInventory().getLeggings() != null)
-                        nw.getInventory().getLeggings().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
-                    if (nw.getInventory().getBoots() != null)
-                        nw.getInventory().getBoots().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
-                }
-            }
-            if (PrisonGame.askType.getOrDefault((Player) sender, 0) == 1) {
-                PrisonGame.setGuard((Player) sender);
-                Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "advancement grant " + sender.getName() + " only prison:guard");
-                Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "advancement grant " + PrisonGame.warden.getName() + " only prison:support");
-                Player nw = (Player) sender;
-                if (nw.getPersistentDataContainer().has(PrisonGame.protspawn, PersistentDataType.INTEGER)) {
-                    if (nw.getInventory().getHelmet() != null)
-                        nw.getInventory().getHelmet().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
-                    if (nw.getInventory().getChestplate() != null)
-                        nw.getInventory().getChestplate().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
-                    if (nw.getInventory().getLeggings() != null)
-                        nw.getInventory().getLeggings().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
-                    if (nw.getInventory().getBoots() != null)
-                        nw.getInventory().getBoots().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
-                }
-            }
-            if (PrisonGame.askType.getOrDefault((Player) sender, 0) == 3) {
-                PrisonGame.setSwat((Player) sender);
-                Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "advancement grant " + sender.getName() + " only prison:guard");
-                Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "advancement grant " + PrisonGame.warden.getName() + " only prison:support");
-                Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "advancement grant " + sender.getName() + " only prison:swat");
-                Player nw = (Player) sender;
-                if (nw.getPersistentDataContainer().has(PrisonGame.protspawn, PersistentDataType.INTEGER)) {
-                    if (nw.getInventory().getHelmet() != null)
-                        nw.getInventory().getHelmet().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
-                    if (nw.getInventory().getChestplate() != null)
-                        nw.getInventory().getChestplate().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
-                    if (nw.getInventory().getLeggings() != null)
-                        nw.getInventory().getLeggings().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
-                    if (nw.getInventory().getBoots() != null)
-                        nw.getInventory().getBoots().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
-                }
-            }
-            PrisonGame.askType.put((Player) sender, 0);
-
-            return true;
-        }
-}
-
-class tgciv implements CommandExecutor {
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        return true;
-    }
-}
-class sfreload implements CommandExecutor {
-
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        HashMap<Player, Location> plhash = new HashMap<>();
-        for (Player p : Bukkit.getOnlinePlayers()) {
-            plhash.put(p, p.getLocation());
-        }
-        new Data(PrisonGame.warden, PrisonGame.active.getName(), true, PrisonGame.swat, plhash).saveData("save.data");
-        for (Player p : Bukkit.getOnlinePlayers()) {
-            Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "advancement grant " + p.getName() + " only prison:reload");
-            p.sendTitle(ChatColor.RED + "LOADING...", "this may take a bit.", 0, Integer.MAX_VALUE, 0);
-            p.sendMessage(ChatColor.RED +  "THE SERVER IS RELOADING!");
-            p.sendMessage(ChatColor.GREEN +  "DO NOT SPAM /WARDEN -" + ChatColor.ITALIC + "YOU WILL BE WARNED!");
-            p.sendMessage(ChatColor.GRAY + "the server will make an attempt to save your items and roles, however this may not work.");
-            p.setGameMode(GameMode.ADVENTURE);
-            p.addPotionEffect(PotionEffectType.DARKNESS.createEffect(Integer.MAX_VALUE, 255));
-            p.addPotionEffect(PotionEffectType.WEAKNESS.createEffect(Integer.MAX_VALUE, 255));
-            p.teleport(new Location(Bukkit.getWorld("world"), -2062, -50, 1945));
-        }
-        Bukkit.getServer().getLogger().log(Level.INFO, "Data Saved");
-        Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "rl confirm");
-        return true;
-    }
-}
-
-class TeamChat implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        if (!Bukkit.getPlayer(sender.getName()).getPersistentDataContainer().has(PrisonGame.muted, PersistentDataType.INTEGER)) {
-            String msg = String.join(" ", args);
-            if (PrisonGame.type.get((Player) sender) == 0) {
-                for (Player p : Bukkit.getOnlinePlayers()) {
-                    if (PrisonGame.type.get(p) == 0) {
-                        p.sendMessage(ChatColor.GRAY + "[" + ChatColor.GOLD + "PRISONER CHAT" + ChatColor.GRAY + "] " + ChatColor.WHITE + sender.getName() + ": " + FilteredWords.filtermsg(msg));
-                    }
-                }
-            }
-            if (PrisonGame.type.get((Player) sender) != 0) {
-                for (Player p : Bukkit.getOnlinePlayers()) {
-                    if (PrisonGame.type.get(p) != 0) {
-                        p.sendMessage(ChatColor.GRAY + "[" + ChatColor.BLUE + "GUARD CHAT" + ChatColor.GRAY + "] " + ChatColor.WHITE + sender.getName() + ": " + FilteredWords.filtermsg(msg));
-                    }
-                }
-            }
-        }
-        return true;
-    }
-}
-
-class settings implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        sender.sendMessage("hello i am bertrude");
-        if (sender instanceof Player) {
-            Player p = (Player) sender;
-            p.playSound(p, Sound.BLOCK_NOTE_BLOCK_BASEDRUM, 1, 1);
-            Inventory inv = Bukkit.createInventory(null, 9, "bertrude");
-            inv.addItem(PrisonGame.createGuiItem(Material.PLAYER_HEAD, ChatColor.BLUE + "old tab", ChatColor.GRAY + "sets tab to the default minecraft one, if you're boring."));
-            inv.addItem(PrisonGame.createGuiItem(Material.POTION, ChatColor.LIGHT_PURPLE + "epic bertude night vision", ChatColor.GRAY + "gives you night vision i think"));
-            inv.addItem(PrisonGame.createGuiItem(Material.GRAY_STAINED_GLASS, ChatColor.LIGHT_PURPLE + "no warden spaces", ChatColor.GRAY + "disables/enables the spaces on the warden's messages"));
-            p.openInventory(inv);
-        }
-        return true;
-    }
-}
-class BuilderCMD implements CommandExecutor {
-
-    // This method is called, when somebody uses our command
-    @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        if (sender instanceof  Player) {
-            Player p = (Player) sender;
-            if (p.hasPermission("minecraft.command.gamemode")) {
-                Inventory inv = Bukkit.createInventory(null, 9 * 6, "BUILDER ZONE");
-                inv.addItem(PrisonGame.createGuiItem(Material.ENDER_PEARL, ChatColor.YELLOW + "Fortress [CMD]", "tp 0 0 0"));
-                inv.addItem(PrisonGame.createGuiItem(Material.ENDER_PEARL, ChatColor.YELLOW + "HyperTech [CMD]", "tp 0 0 -1000"));
-                inv.addItem(PrisonGame.createGuiItem(Material.ENDER_PEARL, ChatColor.YELLOW + "The End [CMD]", "mvtp endprison", ChatColor.RED + "DO /mvtp world TO GET OUT!"));
-                inv.addItem(PrisonGame.createGuiItem(Material.ENDER_PEARL, ChatColor.YELLOW + "Train [CMD]", "tp 0 0 1000"));
-                inv.addItem(PrisonGame.createGuiItem(Material.ENDER_PEARL, ChatColor.YELLOW + "Gladiator [CMD]", "tp -2000 0 2000"));
-                inv.addItem(PrisonGame.createGuiItem(Material.ENDER_PEARL, ChatColor.YELLOW + "Island [CMD]", "tp 2000 0 -2000"));
-                inv.addItem(PrisonGame.createGuiItem(Material.ENDER_PEARL, ChatColor.YELLOW + "Santa's Workshop [CMD]", "tp 2000 0 2000"));
-                inv.addItem(PrisonGame.createGuiItem(Material.ENDER_PEARL, ChatColor.YELLOW + "Volcano [CMD]", "tp -2000 0 -2000"));
-                inv.addItem(PrisonGame.createGuiItem(Material.PLAYER_HEAD, ChatColor.YELLOW + "Teleport All Creative [CMD]", "tp @a[gamemode=creative] @s"));
-                inv.addItem(PrisonGame.createGuiItem(Material.PLAYER_HEAD, ChatColor.YELLOW + "Get Coords [CUSTOM]", "gets you the coords required for me to put in the plugin"));
-                p.openInventory(inv);
-            } else {
-                p.sendMessage("You aren't a builder!");
-            }
-        }
-        return true;
-    }
-}
-
