+++ src/main/java/prisongame/prisongame/commands/AcceptCommand.java
@@ -1,0 +1,70 @@
+package prisongame.prisongame.commands;
+
+import org.bukkit.Bukkit;
+import org.bukkit.command.Command;
+import org.bukkit.command.CommandExecutor;
+import org.bukkit.command.CommandSender;
+import org.bukkit.enchantments.Enchantment;
+import org.bukkit.entity.Player;
+import org.bukkit.persistence.PersistentDataType;
+import prisongame.prisongame.PrisonGame;
+
+public class AcceptCommand implements CommandExecutor {
+
+    // This method is called, when somebody uses our command
+    @Override
+    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
+        if (PrisonGame.askType.getOrDefault((Player) sender, 0) == 2) {
+            PrisonGame.setNurse((Player) sender);
+            Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "advancement grant " + sender.getName() + " only prison:guard");
+            Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "advancement grant " + PrisonGame.warden.getName() + " only prison:support");
+            Player nw = (Player) sender;
+            if (nw.getPersistentDataContainer().has(PrisonGame.protspawn, PersistentDataType.INTEGER)) {
+                if (nw.getInventory().getHelmet() != null)
+                    nw.getInventory().getHelmet().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
+                if (nw.getInventory().getChestplate() != null)
+                    nw.getInventory().getChestplate().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
+                if (nw.getInventory().getLeggings() != null)
+                    nw.getInventory().getLeggings().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
+                if (nw.getInventory().getBoots() != null)
+                    nw.getInventory().getBoots().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
+            }
+        }
+        if (PrisonGame.askType.getOrDefault((Player) sender, 0) == 1) {
+            PrisonGame.setGuard((Player) sender);
+            Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "advancement grant " + sender.getName() + " only prison:guard");
+            Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "advancement grant " + PrisonGame.warden.getName() + " only prison:support");
+            Player nw = (Player) sender;
+            if (nw.getPersistentDataContainer().has(PrisonGame.protspawn, PersistentDataType.INTEGER)) {
+                if (nw.getInventory().getHelmet() != null)
+                    nw.getInventory().getHelmet().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
+                if (nw.getInventory().getChestplate() != null)
+                    nw.getInventory().getChestplate().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
+                if (nw.getInventory().getLeggings() != null)
+                    nw.getInventory().getLeggings().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
+                if (nw.getInventory().getBoots() != null)
+                    nw.getInventory().getBoots().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
+            }
+        }
+        if (PrisonGame.askType.getOrDefault((Player) sender, 0) == 3) {
+            PrisonGame.setSwat((Player) sender);
+            Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "advancement grant " + sender.getName() + " only prison:guard");
+            Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "advancement grant " + PrisonGame.warden.getName() + " only prison:support");
+            Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "advancement grant " + sender.getName() + " only prison:swat");
+            Player nw = (Player) sender;
+            if (nw.getPersistentDataContainer().has(PrisonGame.protspawn, PersistentDataType.INTEGER)) {
+                if (nw.getInventory().getHelmet() != null)
+                    nw.getInventory().getHelmet().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
+                if (nw.getInventory().getChestplate() != null)
+                    nw.getInventory().getChestplate().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
+                if (nw.getInventory().getLeggings() != null)
+                    nw.getInventory().getLeggings().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
+                if (nw.getInventory().getBoots() != null)
+                    nw.getInventory().getBoots().addEnchantment(Enchantment.PROTECTION_ENVIRONMENTAL, 1);
+            }
+        }
+        PrisonGame.askType.put((Player) sender, 0);
+
+        return true;
+    }
+}
