+++ src/main/java/prisongame/prisongame/commands/DiscCommand.java
@@ -1,0 +1,16 @@
+package prisongame.prisongame.commands;
+
+import org.bukkit.ChatColor;
+import org.bukkit.command.Command;
+import org.bukkit.command.CommandExecutor;
+import org.bukkit.command.CommandSender;
+
+public class DiscCommand implements CommandExecutor {
+
+    // This method is called, when somebody uses our command
+    @Override
+    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
+        sender.sendMessage(ChatColor.BLUE + "https://discord.gg/GrcHKkFQsv");
+        return true;
+    }
+}
