+++ src/main/java/prisongame/prisongame/commands/warden/PassCommand.java
@@ -1,0 +1,20 @@
+package prisongame.prisongame.commands.warden;
+
+import org.bukkit.Bukkit;
+import org.bukkit.command.Command;
+import org.bukkit.command.CommandExecutor;
+import org.bukkit.command.CommandSender;
+import org.jetbrains.annotations.NotNull;
+import prisongame.prisongame.MyListener;
+import prisongame.prisongame.PrisonGame;
+
+public class PassCommand implements CommandExecutor {
+    @Override
+    public boolean onCommand(@NotNull CommandSender sender, @NotNull Command command, @NotNull String label, @NotNull String[] args) {
+        MyListener.playerJoin(PrisonGame.warden, false);
+        PrisonGame.warden = null;
+        Bukkit.getPlayer(args[0]).performCommand("warden");
+
+        return true;
+    }
+}
